+++ lending/src/sdk/sdk.ts
@@ -68,9 +68,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "3.5.0";
-    genVersion = "2.143.2";
-    userAgent = "speakeasy-sdk/typescript 3.5.0 2.143.2 3.0.0 @codat/lending";
+    sdkVersion = "3.5.1";
+    genVersion = "2.144.7";
+    userAgent = "speakeasy-sdk/typescript 3.5.1 2.144.7 3.0.0 @codat/lending";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);