+++ lending/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [typescript v3.5.0] lending
 ### Releases
 - [NPM v3.5.0] https://www.npmjs.com/package/@codat/lending/v/3.5.0 - lending
+
+## 2023-10-03 00:05:02
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.5.1] lending
+### Releases
+- [NPM v3.5.1] https://www.npmjs.com/package/@codat/lending/v/3.5.1 - lending